Create mzkit config file from defaults

Create mzkit config file from defaults

mzkit_config_file_from_default(
  default_config_path,
  methodId,
  export_file_path,
  rt_alignment_file_override = "",
  spline_ridge_penalty = 20,
  is_alignment = TRUE,
  is_splitting = FALSE,
  is_coelution = TRUE,
  is_search = TRUE,
  is_qc = TRUE,
  is_eda = TRUE,
  reference_sample_override = NULL,
  is_exclude_isotopes = FALSE
)

Arguments

default_config_path

path to a default config file

methodId

mass spectrometry method being run (e.g., M001A)

export_file_path

path to new config file

rt_alignment_file_override

path to an alignment file or an empty string to ignore

spline_ridge_penalty

spline_ridge_penalty Ridge penalty on sample-level splines which model deviations between observed retention times and the consensus rt of a compound.

is_alignment

should alignment be performed?

is_splitting

should splitting be corrected?

is_coelution

should coelutions be found?

is_search

should search be performed?

is_qc

should QC be performed?

is_eda

should EDA be performed?

reference_sample_override

NULL to skip or the name of a reference sample

is_exclude_isotopes

should isotopologues be excluded?

Value

writes a json config to export_file_path