Run Quahog

Run the quahog pipeline directly from R with a configurable config file.

run_quahog(
  data_folder,
  output_folder = "/tmp/metabolomics_test",
  methodId,
  config_args = list(),
  path_to_mass_spec = "~/mass_spec",
  conda_env,
  conda_path,
  background = FALSE
)

Arguments

data_folder

directory containing sample mzXML/mzML files.

output_folder

directory where output files should be saved.

methodId

mass spectrometry method being run (e.g., M001A)

config_args

list of config arguments to change from the method-specific defaults

path_to_mass_spec

path to mass_spec github repo

conda_env

conda environment to use when running quahog

conda_path

path to conda directory

background

run process in the background?

Value

a proc object for the pipeline running in the background

Details

This is only meant as a convenience function for running mzkit with an appropriate config during development. The primary methods for running quahog is via Docker.